moved files in metplus/util/config to metplus/util
