+++ metplus/util/__init__.py
@@ -1,6 +1,6 @@
+from .config_metplus import *
 from .time_util import *
 from .met_util import *
 from .metplus_check import *
-from .config.config_metplus import *
-from .config.string_template_substitution import *
+from .string_template_substitution import *
 from .feature_util import *
